Version 1.0 - only PvP, no draw
Well, it's a tic-tac-toe game. There's a menu, where you can choose PvP (Player vs AI coming...), and than you play tic-tac-toe. The software check's if one of the player put 5 symbols in a row, and the shows the winner. If you manage to fill all the fields, it's a draw, but nothing happens.